+++ CHANGELOG.md
@@ -1,5 +1,13 @@
 # faupi changelog
 
+## v0.2.0
+
+### Features
+- Response delay
+
+### Changes
+- Add default values to some specification file fields
+
 ## v0.1.0 - Initial Release
 
 ### Features